test without in buffers

+++ src/engine/auxdata/degree_data.hpp
@@ -43,8 +43,8 @@ namespace graphchi {
     
     
     struct degree {
-        int indegree;
-        int outdegree;
+        size_t indegree;
+        size_t outdegree;
     };
